Changeset (#302)
* feat(frontend): enable decentralization mode toggle while in library mode (#275)


* chore(frontend): v0.7.1

* feat(frontend): new links across the frontend (#276)

* feat(frontend): new links across the frontend

* chore(frontend): add landing page link to the footer

* chore(frontend): add 'about' external landing page link to the navbar

---------


* fix(ui): disable click on tokens while loading (#278)

* feat(frontend,ui): typography review (#277)

* feat(frontend): new links across the frontend

* chore(frontend): add landing page link to the footer

* chore(frontend): add 'about' external landing page link to the navbar

* feat(ui): new typography styles for both markdown and standard typography

* feat(frontend): adapt frontend to new ui config

* feat(ui): new line height for headings

* chore(ui): slightly tweak headings letter spacing

* fix(ui): make sure switzer font family is applied to markdown component's titles

* feat(frontend): reduce carrot logo width in footer

---------


* chore(ui): v0.47.0

* feat(frontend): remove forced decentralization mode setting in library mode

* chore(frontend): v0.7.2

* feat(frontend): new favicons (#279)

* feat(frontend): new favicons

* feat(frontend): new og thumb image

---------


* chore: upgrade dependencies (#280)


* chore(sdk): v1.27.4

* chore(ui): v0.47.1

* chore(react): v0.47.1

* chore(frontend): v0.7.3

* chore: explicity ts type imports

* chore(sdk): v1.27.5

* chore(ui): v0.47.2

* chore(react): v0.47.2

* chore(frontend): v0.7.4

* feat(ui): update step card title sizes

* chore(ui): v0.48.0

* fix(ui): fix select types (#281)

* fix(ui): fix select types

* fix(frontrend,ui): fix select typing

---------


* chore(ui): v0.48.1

* chore(frontend): v0.7.5

* chore(ui): fix types on select component

* chore(ui): v0.48.2

* feat(ui): avoid dorcing text size reduction for headings on mobile

* chore(ui): v0.48.3

* Fetch latest KPI tokens (#282)

* feat(sdk): add fetchLatestKPITokenAddresses fetcher

* fix(frontend): use new fetcher in the useLatestKPITokens hook

* fix(sdk): sort the KPI token by asc

* fix(frontend): reverse the latest KPI tokens array

* refactor: replace slice with cast

* refactor(sdk): move fetchers validation to common functions

* feat(ui): typography update (#283)


* fix(frontend): remove minimum height on hero

* fix(frontend): fix video not visible on safari in hero

* Add default block explorers for supported chains (#284)

* fix(frontend): add default block explorers for supported chains

* chore: upgrade dependencies

* chore: rollback to react-redux v8.0.7

* fix(ui): allow non serializable select values (#285)

* fix(ui): allow non serializable select values

* chore(ui): add use callback

* fix(ui): remove extra check

---------


* chore(ui): v0.48.4

* chore(sdk): format currency amount instead of token amount

* chore(sdk): v1.28.0

* fix(ui): fix sans body text typography config (#286)


* fix(ui): typography heading text size fix (#287)

* fix(ui): fix sans body text typography config

* chore(ui): v0.48.5

* fix(ui): fix heading typography config

---------


* chore(ui): v0.48.6

* fix(frontend): use ankr rpc for gnosis (#288)

* Collateral recover transaction (#289)

* feat(frontend): add ERC20 recover transaction types

* feat(react): add recover ERC20 transaction types

* feat(sdk): update KPI token abi

* chore(sdk): v1.28.1

* chore(sdk): v0.48.0

* chore(frontend): v0.8.0

* Upgrade dependencies (#290)

* chore: upgrade dependencies

* fix: downgrade eslint-config-turbo to v1.10.3

* fix: lock eslint-config-turbo version

* chore(sdk): v1.28.2

* chore(sdk): v1.28.2

* chore(ui): v0.48.7

* chore(react): v0.48.1

* chore(frontend): v0.8.1

* chore: upgrade deps (#291)

* chore: upgrade patched dependencies

* chore: upgrade minor deps

* chore: upgrade remaining dependencies

* chore: update peer dependencies and shared dependencies json

---------


* feat(react): avoid passing t function to federated components (#292)


* chore(react): v0.49.0

* chore(frontend): v0.9.0

* fix(subgraph): fix upgrade templates issue on subgraph (#294)

* chore(subgraph): update abis

* fix(subgraph): fix template upgrades overwriting previous template version

---------


* Revert "feat(react): avoid passing t function to federated components (#292)" (#295)

This reverts commit a1d2d3cbd5b9ea484c6261e462361c0054da5066.


* chore(react): v0.50.0

* chore(frontend): v0.10.0

* feat(react): avoid passing t function to federated components (#292) (#296)


* Revert "feat(react): avoid passing t function to federated components (#292) (#296)"

This reverts commit 1e75952e9ffd91c77498c31b2ee1a2f5d6d7dd7a.

* chore: update addresses (#297)

* chore(sdk): update addresses

* chore(sdk): update abis

* chore(subgraph): update gnosis addresses

---------


* chore(sdk): v1.29.0

* chore(ui): v0.48.8

* chore(react): v0.50.1

* chore(frontend): v0.11.0

* feat(react): avoid passing t function to federated components (#292)


* chore(react): v0.50.2

* chore(frontend): v0.11.1

* Revert "feat(react): avoid passing t function to federated components (#292)"

This reverts commit 06ec417feee84059a8b436c610f7d3c62e9d455a.

* chore(react): v0.50.3

* chore(frontend): v0.11.2

* Css issue when loading ERC20 tokens (#293)

* fix(ui): css issue when loading ERC20 tokens

* fix(frontend): replace expired discord link

* fix(frontend): minor css issue

* Improve mobile menu, fix minor issues (#298)

* feat(frontend): improve mobile side menu

* fix(frontend): issues with the campaigns page

* chore: update repo link

* feat(frontend): mp4 source in hero video

* feat: introduce new ipfs pinning proxy (#299)

* feat: introduce new ipfs pinning proxy

* chore(frontend): remove extra if

---------


* Avoid resolving all the KPI tokens at once (#300)

* feat(frontend): avoid resolving all the kpi tokens at once

* feat(frontend): responsiveness and style fixes and hide template filters on campaigns page

* chore(frontend): ordering change in fathomrc

* fix(frontend): avoid grid flashing when tokens have been fetched but not yet transformed into array

---------


* chore: upgrade patched dependencies

* chore: upgrade minor dependencies and update storybook

* chore: upgrade prettier and eslint

* chore: update typescript eslint plugin

* chore: upgrade more dependencies

* chore: add changeset CI (#301)

* chore: add changeset CI

* docs: add CHANGELOG.md file with changeset info

* chore: remove .yarn folder

---------


* chore: add first changeset

---------

+++ packages/sdk/src/abis/kpi-token.ts
@@ -4,9 +4,9 @@ export default [
         name: "creationTimestamp",
         outputs: [
             {
-                internalType: "uint64",
+                internalType: "uint256",
                 name: "",
-                type: "uint64",
+                type: "uint256",
             },
         ],
         stateMutability: "view",
@@ -43,9 +43,9 @@ export default [
         name: "expiration",
         outputs: [
             {
-                internalType: "uint64",
+                internalType: "uint256",
                 name: "",
-                type: "uint64",
+                type: "uint256",
             },
         ],
         stateMutability: "view",
